Harden external object sync edge cases
