+++ ui/src/lib/external-objects.test.ts
@@ -1,4 +1,10 @@
-import { describe, expect, it } from "vitest";
+import { afterEach, describe, expect, it, vi } from "vitest";
+import type { ExternalObjectSummary } from "@paperclipai/shared";
+import { externalObjectsApi } from "../api/externalObjects";
+import {
+  EXTERNAL_OBJECT_SUMMARY_BATCH_SIZE,
+  fetchIssueExternalObjectSummariesInBatches,
+} from "../hooks/useIssueExternalObjects";
 import {
   dominantExternalObjectTone,
   externalObjectCategoryLabel,
@@ -13,6 +19,20 @@ import {
   sortExternalObjectsBySeverity,
 } from "./external-objects";
 import { normalizeExternalObjectHref } from "./external-object-href";
+
+vi.mock("../api/externalObjects", () => ({
+  externalObjectsApi: {
+    getIssueSummaries: vi.fn(),
+  },
+}));
+
+const emptySummary: ExternalObjectSummary = {
+  total: 0,
+  byStatusCategory: {},
+  byLiveness: {},
+  highestSeverity: "neutral",
+  objects: [],
+};
 
 describe("normalizeExternalObjectHref", () => {
   it("lowercases the host (case preserving the path) and drops query/fragment", () => {
@@ -32,6 +52,33 @@ describe("normalizeExternalObjectHref", () => {
 
   it("defaults pathless URLs to /", () => {
     expect(normalizeExternalObjectHref("https://example.com")).toBe("https://example.com/");
+  });
+});
+
+describe("fetchIssueExternalObjectSummariesInBatches", () => {
+  afterEach(() => {
+    vi.mocked(externalObjectsApi.getIssueSummaries).mockReset();
+  });
+
+  it("chunks bulk summary requests below the server validation cap and merges results", async () => {
+    const issueIds = Array.from(
+      { length: EXTERNAL_OBJECT_SUMMARY_BATCH_SIZE * 2 + 3 },
+      (_entry, index) => `issue-${index}`,
+    );
+    vi.mocked(externalObjectsApi.getIssueSummaries).mockImplementation(async (_companyId, ids) => ({
+      summaries: Object.fromEntries(ids.map((id) => [id, { ...emptySummary, total: id.endsWith("-0") ? 1 : 0 }])),
+    }));
+
+    const result = await fetchIssueExternalObjectSummariesInBatches("company-1", issueIds);
+
+    expect(externalObjectsApi.getIssueSummaries).toHaveBeenCalledTimes(3);
+    expect(vi.mocked(externalObjectsApi.getIssueSummaries).mock.calls.map((call) => call[1].length)).toEqual([
+      EXTERNAL_OBJECT_SUMMARY_BATCH_SIZE,
+      EXTERNAL_OBJECT_SUMMARY_BATCH_SIZE,
+      3,
+    ]);
+    expect(Object.keys(result.summaries)).toHaveLength(issueIds.length);
+    expect(result.summaries["issue-0"]?.total).toBe(1);
   });
 });
